Client side completed
Created an interface to welcome the players and allow them to decide what trivia game they would like to play, what server they would play on, and a help option. 
Inside the client file, the client-to-server connection was implemented. First, a socket was created to talk to the server and then step 2 was to connect to the server. The code checks to make sure a proper server address was entered and then checks to see if the server is accepting connections.